// src/types/OperationIds.type.ts
export enum OperationIds {
  LOGIN = 'login',
  FORGOT_PASSWORD = 'forgotPassword',
  RESET_PASSWORD = 'resetPassword',
  COMPANIES = 'companies',
  CREATE_COMPANY = 'createCompany',
  EDIT_COMPANY = 'editCompany',
  DELETE_COMPANY = 'deleteCompany',
  GET_DETAIL_COMPANY = 'getDetailCompany',
  GET_SECTORS = 'getSectors',
  EXAMS = 'getExams',
  GET_USERS = 'getUsers',
  GET_DIAGNOSES = 'getDiagnoses',
  SEND_EMAIL_INSTRUCTION = 'sendEmailInstruction',
  CURRENT_ADMIN = 'currentAdmin',
  GET_USER = 'getUser',
  GET_DATA_REPORT = 'getDataReport',
}