// src/types/StoreModules.ts
export enum StoreModuleNames {
  FEA_DEMO = 'features/demo',
  FEA_LOGIN = 'features/login',
  REQ_LOGIN = 'req/login',
  REQ_FORGOT_PASSWORD = 'req/forgotPassword',
  REQ_RESET_PASSWORD = 'req/resetPassword',
  REQ_CREATE_COMPANY = 'req/createCompany',
  REQ_EDIT_COMPANY = 'req/editCompany',
  REQ_GET_COMPANIES = 'req/companies',
  REQ_GET_EXAMS = 'req/getExams',
  REQ_GET_USERS = 'req/getUsers',
  REQ_GET_DIAGNOSES = 'req/getDiagnoses',
  APP = 'app',
}